requirements.txt
fastapi
uvicorn
httpx
nltk
spacy<3.7.0
pysbd
transformers>=4.30.0
torch>=1.12.0
sentence-transformers>=2.2.2
faiss-cpu
pydantic
pydantic-settings
pytest
PyJWT
redis
black
isort
biopython
python-dotenv
pytest-asyncio
mypy
types-requests
gradio
pandas
fpdf2
pdfplumber
requests
streamlit
beautifulsoup4
chromadb>=0.4.14
duckdb
matplotlib
scikit-learn
huggingface-hub
pdfplumber==0.10.2
fastapi
uvicorn
jinja2>=3.1.0
markupsafe>=2.1.0
PyPDF2>=3.0.1
feedparser>=6.0.0
